common steps improvements
