+++ src/test/java/stepdefinitions/BaseSteps.java
@@ -13,11 +13,12 @@ import java.time.Duration;
 
 public class BaseSteps {
     public static WebDriver driver;
-
+//    public static String baseUrl = "https://hepikorea.pad19.me";
+    public static String baseUrl = "https://hk.crazyz.biz.id";
 
     @Given("user membuka platform HepiKorea")
     public void user_membuka_platform_hepi_korea() {
-        driver.get("https://hepikorea.pad19.me/auth/login");
+        driver.get(baseUrl+"/auth/login");
 
         // Tambahan: pastikan field email muncul
         new WebDriverWait(driver, Duration.ofSeconds(10))
@@ -31,7 +32,12 @@ public class BaseSteps {
             // Jika kamu sudah pakai Selenium 4.6+, tidak perlu lagi setProperty
             // System.setProperty("webdriver.chrome.driver", "path/to/chromedriver.exe");
 
-            driver = new ChromeDriver();
+            // Tambahkan preferensi untuk mengurangi keamanan saat testing
+            var options = new org.openqa.selenium.chrome.ChromeOptions();
+            options.addArguments("--disable-web-security");
+            options.addArguments("--allow-running-insecure-content");
+
+            driver = new ChromeDriver(options);
             driver.manage().window().maximize();
             driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
         }
