+++ src/test/java/stepdefinitions/FinishOrderSteps.java
@@ -1,13 +1,10 @@
 package stepdefinitions;
 
 import io.cucumber.java.en.And;
-import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import org.example.HistoryPage;
 import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
@@ -18,18 +15,9 @@ public class FinishOrderSteps {
     WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(10));
     HistoryPage historyPage;
 
-    @Given("user telah melakukan login sebagai customers")
-    public void user_login_customers() {
-        driver.get("https://hk.crazyz.biz.id/auth/login");
-        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//input[@placeholder='Email']"))).sendKeys("testuser@example.com");
-        driver.findElement(By.xpath("//input[@placeholder='Password']")).sendKeys("testpassword");
-        driver.findElement(By.xpath("//button[contains(text(),'Login')]")).click();
-        wait.until(ExpectedConditions.urlToBe("https://hk.crazyz.biz.id/"));
-    }
-
     @And("user membuka halaman sent")
     public void user_membuka_halaman_sent() {
-        driver.get("https://hk.crazyz.biz.id/order/sent");
+        driver.get(BaseSteps.baseUrl+"/order/sent");
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//button[contains(text(),'Sent')]")));
         historyPage = new HistoryPage(driver);
     }
